+++ frontend/packages/lib-shared/method/formCreate.ts
@@ -1,6 +1,6 @@
 import type { CommonList } from '../models/common';
 import { FieldTypeEnum } from '../enums/formDesignEnum';
-import type { FormCreateField } from '@cordys/web/src/components/business/crm-form-create/types';
+import type { FormCreateField, FormDetail } from '@cordys/web/src/components/business/crm-form-create/types';
 import { formatTimeValue, getCityPath, getIndustryPath } from './index';
 import type { ModuleField } from '../models/customer';
 import { useI18n } from '../hooks/useI18n';
@@ -21,6 +21,14 @@ export const multipleTypes = [FieldTypeEnum.CHECKBOX, FieldTypeEnum.SELECT_MULTI
 export const memberTypes = [FieldTypeEnum.MEMBER, FieldTypeEnum.MEMBER_MULTIPLE];
 export const departmentTypes = [FieldTypeEnum.DEPARTMENT, FieldTypeEnum.DEPARTMENT_MULTIPLE];
 export const singleTypes = [FieldTypeEnum.RADIO, FieldTypeEnum.SELECT];
+export const specialBusinessKeyMap: Record<string, string> = {
+  customerId: 'customerName',
+  contactId: 'contactName',
+  clueId: 'clueName',
+  businessId: 'businessName',
+  contractId: 'contractName',
+  owner: 'ownerName',
+};
 
 export function getRuleType(item: FormCreateField) {
   if (
@@ -88,6 +96,95 @@ export function formatNumberValue(value: string | number, item: FormCreateField)
   return '-';
 }
 
+export function initFieldValue(field: FormCreateField, value: string | number | (string | number)[]) {
+  if (
+    [FieldTypeEnum.DATA_SOURCE, FieldTypeEnum.DATA_SOURCE_MULTIPLE].includes(field.type) &&
+    typeof value === 'string'
+  ) {
+    return value ? [value] : [];
+  }
+  return value;
+}
+
+export function parseModuleFieldValue(item: FormCreateField, fieldValue: string | string[], options?: any[]) {
+  const { t } = useI18n();
+  let value: string | string[] = fieldValue || '';
+  if (options) {
+    // 若字段值是选项值，则取选项值的name
+    if (Array.isArray(fieldValue)) {
+      value = fieldValue.map((e) => {
+        const option = options.find((opt) => opt.id === e);
+        if (option) {
+          return option.name || t('common.optionNotExist');
+        }
+        return t('common.optionNotExist');
+      });
+    } else {
+      value = options.find((e) => e.id === fieldValue)?.name || t('common.optionNotExist');
+    }
+  } else if (item.type === FieldTypeEnum.LOCATION) {
+    const addressArr: string[] = (fieldValue as string).split('-') || [];
+    value = addressArr.length ? `${getCityPath(addressArr[0])}-${addressArr.filter((e, i) => i > 0).join('-')}` : '-';
+  } else if (item.type === FieldTypeEnum.INDUSTRY) {
+    value = fieldValue ? getIndustryPath(fieldValue as string) : '-';
+  } else if (item.type === FieldTypeEnum.INPUT_NUMBER) {
+    value = formatNumberValue(fieldValue as string, item);
+  } else if (item.type === FieldTypeEnum.DATE_TIME) {
+    value = formatTimeValue(fieldValue as string, item.dateType);
+  }
+  if (Array.isArray(value) && item.resourceFieldId) {
+    value = value.join(',');
+  }
+  return value;
+}
+
+export function parseFormDetailValue(item: FormCreateField, form: FormDetail, sourceName?: Ref<string>) {
+  const { t } = useI18n();
+  if (item.businessKey && !item.resourceFieldId) {
+    // 引用数据源字段使用 id 读取数据，而不是 businessKey
+    const options = form.optionMap?.[item.businessKey];
+    // 业务标准字段读取最外层，读取form[item.businessKey]取到 id 值，然后去 options 里取 name
+    let name: string | string[] = '';
+    const value = form[item.businessKey];
+    // 若字段值是选项值，则取选项值的name
+    if (options) {
+      if (Array.isArray(value)) {
+        name = value.map((e) => {
+          const option = options.find((opt) => opt.id === e);
+          if (option) {
+            return option.name || t('common.optionNotExist');
+          }
+          return t('common.optionNotExist');
+        });
+      } else {
+        name = options.find((e) => e.id === value)?.name || t('common.optionNotExist');
+      }
+    }
+    if (item.type === FieldTypeEnum.DATE_TIME) {
+      return formatTimeValue(name || form[item.businessKey], item.dateType);
+    }
+    if (item.type === FieldTypeEnum.INPUT_NUMBER) {
+      return formatNumberValue(name || form[item.businessKey], item);
+    }
+    if (item.type === FieldTypeEnum.ATTACHMENT) {
+      return form.attachmentMap?.[item.businessKey] || [];
+    }
+    if (item.businessKey === 'name' && sourceName) {
+      sourceName.value = name || form[item.businessKey];
+    }
+    return name || form[item.businessKey];
+  }
+  const options = form.optionMap?.[item.id];
+  // 其他的字段读取moduleFields
+  const field = form.moduleFields?.find((moduleField: ModuleField) => moduleField.fieldId === item.id);
+  if (item.type === FieldTypeEnum.ATTACHMENT) {
+    return form.attachmentMap?.[item.id] || [];
+  }
+  if (field) {
+    return parseModuleFieldValue(item, field.fieldValue, options);
+  }
+}
+
 /**
  * 表单配置表格回显数据
  */
@@ -124,13 +221,24 @@ export function transformData({
     } else if (field.type === FieldTypeEnum.DATE_TIME) {
       timeFieldIds.push(fieldId);
     } else if ([FieldTypeEnum.SUB_PRICE, FieldTypeEnum.SUB_PRODUCT].includes(field.type) && needParseSubTable) {
-      subTableFieldInfo = transformData({
-        item: item[fieldId],
-        fields: field.subFields || [],
-        originalData,
-        excludeFieldIds,
+      field.subFields?.forEach((subField) => {
+        item[fieldId]?.forEach((subItem: Record<string, any>) => {
+          if (subField.resourceFieldId) {
+            subItem[subField.id] = parseModuleFieldValue(
+              subField,
+              subItem[subField.businessKey || subField.id],
+              originalData?.optionMap?.[subField.id]
+            );
+          } else {
+            // 不使用业务 key，直接使用字段 id 去取值
+            subItem[subField.id] = parseModuleFieldValue(
+              subField,
+              subItem[subField.businessKey || subField.id],
+              originalData?.optionMap?.[subField.businessKey || subField.id]
+            );
+          }
+        });
       });
-      console.log('subTableFieldInfo', subTableFieldInfo);
     }
     if (field.businessKey) {
       const fieldId = field.businessKey;
@@ -182,6 +290,9 @@ export function transformData({
         if (fieldId === 'owner') {
           businessFieldAttr.ownerId = item.owner;
         }
+      } else if (specialBusinessKeyMap[fieldId]) {
+        // 处理特殊业务 key 映射关系
+        businessFieldAttr[fieldId] = item[specialBusinessKeyMap[fieldId]];
       }
     }
   });
